Add files via upload
The python format that I have been committing to thus far has been good, but lately it has gotten quite messy due to the increasing size of the project and as a result, it's becoming harder to distinguish each individual part of the project from each other. I find this Jupyter Notebook format a lot neater. It also allows me to presents graphs and tables in the future, if I ever want to.